Add files via upload
The object of the rock-paper-scissor python project is to build a game for a single player that plays with a computer, anywhere, and anytime. This project is base on the rules that:
- rock blunts scissors so rock wins
- scissors cut the paper so scissors win
- paper cover rock so paper wins
This project is build using tkinter, random modules, and the basic concept of python.
In this python project, players have to choose any one from rock, paper, and scissors. Then click on the play button will show the result of the game.